requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
mypy-extensions==0.4.1    # via mypy
mypy==0.701
numpy==1.16.2             # via pandas
pandas==0.22.0
ply==3.11
python-bidi==0.4.0
python-dateutil==2.8.0    # via pandas
pytz==2018.9              # via pandas
pyyaml==5.1
six==1.12.0               # via python-bidi, python-dateutil
typed-ast==1.3.5          # via mypy
xlrd==1.2.0
